+++ src/test/kotlin/com/terraformation/backend/db/DatabaseBackedTest.kt
@@ -3675,7 +3675,7 @@ abstract class DatabaseBackedTest {
       row: ReportProjectIndicatorsRow = ReportProjectIndicatorsRow(),
       reportId: ReportId = row.reportId ?: inserted.reportId,
       indicatorId: ProjectIndicatorId = row.projectIndicatorId ?: inserted.projectIndicatorId,
-      value: Int? = row.value,
+      value: BigDecimal? = row.value,
       projectsComments: String? = row.projectsComments,
       progressNotes: String? = row.progressNotes,
       status: ReportIndicatorStatus? = row.statusId,
@@ -3701,7 +3701,7 @@ abstract class DatabaseBackedTest {
       row: ReportCommonIndicatorsRow = ReportCommonIndicatorsRow(),
       reportId: ReportId = row.reportId ?: inserted.reportId,
       indicatorId: CommonIndicatorId = row.commonIndicatorId ?: inserted.commonIndicatorId,
-      value: Int? = row.value,
+      value: BigDecimal? = row.value,
       projectsComments: String? = row.projectsComments,
       progressNotes: String? = row.progressNotes,
       status: ReportIndicatorStatus? = row.statusId,
@@ -3728,9 +3728,9 @@ abstract class DatabaseBackedTest {
       reportId: ReportId = row.reportId ?: inserted.reportId,
       indicator: AutoCalculatedIndicator =
           row.autoCalculatedIndicatorId ?: AutoCalculatedIndicator.SeedsCollected,
-      systemValue: Int? = row.systemValue,
+      systemValue: BigDecimal? = row.systemValue,
       systemTime: Instant? = row.systemTime ?: systemValue?.let { Instant.EPOCH },
-      overrideValue: Int? = row.overrideValue,
+      overrideValue: BigDecimal? = row.overrideValue,
       projectsComments: String? = row.projectsComments,
       progressNotes: String? = row.progressNotes,
       status: ReportIndicatorStatus? = row.statusId,
@@ -3760,7 +3760,7 @@ abstract class DatabaseBackedTest {
       projectIndicatorId: ProjectIndicatorId =
           row.projectIndicatorId ?: inserted.projectIndicatorId,
       year: Int = row.year ?: 1970,
-      target: Int? = row.target,
+      target: BigDecimal? = row.target,
   ) {
     val rowWithDefaults =
         row.copy(
@@ -3778,7 +3778,7 @@ abstract class DatabaseBackedTest {
       projectId: ProjectId = row.projectId ?: inserted.projectId,
       commonIndicatorId: CommonIndicatorId = row.commonIndicatorId ?: inserted.commonIndicatorId,
       year: Int = row.year ?: 1970,
-      target: Int? = row.target,
+      target: BigDecimal? = row.target,
   ) {
     val rowWithDefaults =
         row.copy(
@@ -3797,7 +3797,7 @@ abstract class DatabaseBackedTest {
       indicator: AutoCalculatedIndicator =
           row.autoCalculatedIndicatorId ?: AutoCalculatedIndicator.SeedsCollected,
       year: Int = row.year ?: 1970,
-      target: Int? = row.target,
+      target: BigDecimal? = row.target,
   ) {
     val rowWithDefaults =
         row.copy(
@@ -3816,7 +3816,7 @@ abstract class DatabaseBackedTest {
       projectIndicatorId: ProjectIndicatorId =
           row.projectIndicatorId ?: inserted.projectIndicatorId,
       year: Int = row.year ?: 2025,
-      target: Int? = row.target,
+      target: BigDecimal? = row.target,
   ) {
     val rowWithDefaults =
         row.copy(
@@ -3834,7 +3834,7 @@ abstract class DatabaseBackedTest {
       projectId: ProjectId = row.projectId ?: inserted.projectId,
       commonIndicatorId: CommonIndicatorId = row.commonIndicatorId ?: inserted.commonIndicatorId,
       year: Int = row.year ?: 2025,
-      target: Int? = row.target,
+      target: BigDecimal? = row.target,
   ) {
     val rowWithDefaults =
         row.copy(
@@ -3854,7 +3854,7 @@ abstract class DatabaseBackedTest {
       indicator: AutoCalculatedIndicator =
           row.autoCalculatedIndicatorId ?: AutoCalculatedIndicator.SeedsCollected,
       year: Int = row.year ?: 2025,
-      target: Int? = row.target,
+      target: BigDecimal? = row.target,
   ) {
     val rowWithDefaults =
         row.copy(
@@ -4061,7 +4061,7 @@ abstract class DatabaseBackedTest {
       row: PublishedReportProjectIndicatorsRow = PublishedReportProjectIndicatorsRow(),
       reportId: ReportId = row.reportId ?: inserted.reportId,
       indicatorId: ProjectIndicatorId = row.projectIndicatorId ?: inserted.projectIndicatorId,
-      value: Int? = row.value,
+      value: BigDecimal? = row.value,
       projectsComments: String? = row.projectsComments,
       progressNotes: String? = row.progressNotes,
       status: ReportIndicatorStatus? = row.statusId,
@@ -4083,7 +4083,7 @@ abstract class DatabaseBackedTest {
       row: PublishedReportCommonIndicatorsRow = PublishedReportCommonIndicatorsRow(),
       reportId: ReportId = row.reportId ?: inserted.reportId,
       indicatorId: CommonIndicatorId = row.commonIndicatorId ?: inserted.commonIndicatorId,
-      value: Int? = row.value,
+      value: BigDecimal? = row.value,
       projectsComments: String? = row.projectsComments,
       progressNotes: String? = row.progressNotes,
       status: ReportIndicatorStatus? = row.statusId,
@@ -4107,7 +4107,7 @@ abstract class DatabaseBackedTest {
       reportId: ReportId = row.reportId ?: inserted.reportId,
       indicator: AutoCalculatedIndicator =
           row.autoCalculatedIndicatorId ?: AutoCalculatedIndicator.Seedlings,
-      value: Int? = row.value,
+      value: BigDecimal? = row.value,
       projectsComments: String? = row.projectsComments,
       progressNotes: String? = row.progressNotes,
       status: ReportIndicatorStatus? = row.statusId,
